divided prepareTokensData function to two functions
